+++ packages/react-templates/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.39-alpha.3](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.1.39-alpha.2...@selfcommunity/react-templates@0.1.39-alpha.3) (2022-09-01)
+
+**Note:** Version bump only for package @selfcommunity/react-templates
+
+
+
+
+
+## [0.1.39-alpha.2](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.1.39-alpha.1...@selfcommunity/react-templates@0.1.39-alpha.2) (2022-09-01)
+
+**Note:** Version bump only for package @selfcommunity/react-templates
+
+
+
+
+
+## [0.1.39-alpha.1](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.1.39-alpha.0...@selfcommunity/react-templates@0.1.39-alpha.1) (2022-08-31)
+
+**Note:** Version bump only for package @selfcommunity/react-templates
+
+
+
+
+
+## [0.1.39-alpha.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.1.38...@selfcommunity/react-templates@0.1.39-alpha.0) (2022-08-31)
+
+**Note:** Version bump only for package @selfcommunity/react-templates
+
+
+
+
+
 ## [0.1.38](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.1.38-alpha.8...@selfcommunity/react-templates@0.1.38) (2022-08-31)
 
 **Note:** Version bump only for package @selfcommunity/react-templates